pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "giga-mcp-server"
version = "0.1.0"
description = "MCP server bridging WhatsApp group messages to JIRA Kanban board stories"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.8.0,<2.0",
    "atlassian-python-api>=3.41",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "structlog>=24.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.40"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock",
    "respx>=0.22",
    "ruff>=0.8",
]

[project.scripts]
giga-mcp-server = "giga_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/giga_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
